Removed redunt data class
