+++ app/src/main/java/com/trelp/aag2020/data/Movie.kt
@@ -5,13 +5,16 @@ import androidx.annotation.DrawableRes
 data class Movie(
     val id: Int,
     @DrawableRes val posterSmall: Int,
+    @DrawableRes val poster: Int,
     val ageLimit: String,
     val title: String,
     val tags: String,
     val rating: Float,
     val reviewCount: Int,
     val isLike: Boolean,
-    val duration: Int
+    val duration: Int,
+    val overview: String,
+    val actors: List<Actor>
 )
 
 data class MovieDetails(
