+++ app/src/main/java/com/trelp/aag2020/ui/details/ActorAdapter.kt
@@ -8,7 +8,7 @@ import com.trelp.aag2020.ui.common.utils.inflater
 
 class ActorAdapter : RecyclerView.Adapter<ActorAdapter.ActorHolder>() {
 
-    private var data = listOf<Actor>()
+    private val data = mutableListOf<Actor>()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ActorHolder {
         return ActorHolder(
@@ -23,7 +23,7 @@ class ActorAdapter : RecyclerView.Adapter<ActorAdapter.ActorHolder>() {
     override fun getItemCount() = data.size
 
     fun setupData(newData: List<Actor>) {
-        data = newData
+        data.plusAssign(newData)
         notifyDataSetChanged()
     }
